requirements.txt
attrdict==2.0.1
einops==0.4.1
matplotlib==3.5.2
numpy==1.21.6
pillow==10.2.0
pyyaml==6.0.1
scikit_learn==1.0.2
torch==1.11.0
torchvision==0.13
tqdm==4.64.0
